Big cleanup
The project is now separated into the library and the binary to avoid
sharing dependencies. The dependency tree has also been cleaned up
from non-useful dependencies, and the ones that are left have been checked
manually, and their default features disabled.

Basically, a much cleaner room to work with, and less dependency creep.